Separate frontend and classifier
Separating these will make it easier to create a standalone API application.
